requirements.txt
transformers
pyyaml
black
spacy
spacy-langdetect
